+++ QCloudAPI_SDK/QCloudAPI_SDK/Module/Tbaas.cs
@@ -1,0 +1,10 @@
+namespace QCloudAPI_SDK.Module
+{
+    public class Tbaas : Base
+    {
+        public Tbaas()
+        {
+            serverHost = "tbaas.api.qcloud.com";
+        }
+    }
+}